Refactor : config패키지 생성 및 경로 변경
