+++ Flotting-backend/src/main/java/com/flotting/config/QueryDslConfig.java
@@ -1,4 +1,4 @@
-package com.flotting;
+package com.flotting.config;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import jakarta.persistence.EntityManager;
