+++ Flotting-backend/src/main/java/com/flotting/config/TokenUser.java
@@ -1,4 +1,4 @@
-package com.flotting.api.user.controller;
+package com.flotting.config;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -12,4 +12,8 @@ public class TokenUser {
     private String id;
 
     private String name;
+
+    public static TokenUser defaultUser() {
+        return new TokenUser("id", "defaultUser");
+    }
 }
